Modified qhost to utilize the qstat_reduce header, and to respect qstat_reduce, running the old code if qstat_reduce is not yet, otherwise pulling from the cached xml file.

+++ qhost.php
@@ -11,6 +11,7 @@
 <?php
 $owner  = $_GET['owner'];
 echo "<body><table align=center width=95% border=\"1\" cellpadding=\"0\" cellspacing=\"0\"><tbody>";
+include("header.php");
 echo "<tr><td><h1>PHPQstat</h1></td></tr>
       <tr><td CLASS=\"bottom\" align=center><a href='index.php'>Home</a> *  <a href=\"qhost.php?owner=$owner\">Hosts status</a> *  <a href=\"qstat.php?owner=$owner\">Queue status</a> * <a href=\"qstat_user.php?owner=$owner\">Jobs status ($owner)</a> * <a href=\"about.php?owner=$owner\">About PHPQstat</a></td></tr>";
 ?>
@@ -32,25 +33,31 @@ echo "<tr><td><h1>PHPQstat</h1></td></tr>
                 <td>swap_used</td>
                 </tr>
 <?php
-$password_length = 20;
+if ($qstat_reduce != "yes") {
 
-function make_seed() {
-  list($usec, $sec) = explode(' ', microtime());
-  return (float) $sec + ((float) $usec * 100000);
+	$password_length = 20;
+
+	function make_seed() {
+	  list($usec, $sec) = explode(' ', microtime());
+	  return (float) $sec + ((float) $usec * 100000);
+	}
+
+	srand(make_seed());
+
+	$alfa = "1234567890qwertyuiopasdfghjklzxcvbnmQWERTYUIOPASDFGHJKLZXCVBNM";
+	$token = "";
+	for($i = 0; $i < $password_length; $i ++) {
+	  $token .= $alfa[rand(0, strlen($alfa))];
+	}
+
+	$out = exec("./qhostout /tmp/$token.xml");
+
+	//printf("System Output: $out\n"); 
+	$qhost = simplexml_load_file("/tmp/$token.xml");
+} else {
+	$qhost = simplexml_load_file("/tmp/qhost.xml");
 }
 
-srand(make_seed());
-
-$alfa = "1234567890qwertyuiopasdfghjklzxcvbnmQWERTYUIOPASDFGHJKLZXCVBNM";
-$token = "";
-for($i = 0; $i < $password_length; $i ++) {
-  $token .= $alfa[rand(0, strlen($alfa))];
-}
-
-$out = exec("./qhostout /tmp/$token.xml");
-
-//printf("System Output: $out\n"); 
-$qhost = simplexml_load_file("/tmp/$token.xml");
 $i=0;
 foreach ($qhost->host as $host) {
 	echo "<tr>";
@@ -63,8 +70,10 @@ foreach ($qhost->host as $host) {
 	$i++;
 }
 
+if ($qstat_reduce != "yes") {
+	exec("rm /tmp/$token.xml");
+}
 
-exec("rm /tmp/$token.xml");
 ?>
 
 	  </tbody>
